style: fix pre-existing clippy warnings to enable the -D warnings gate
Clears the 56 clippy warnings that accumulated while clippy was never enforced in CI, so the new rust job can gate on -D warnings. Mostly mechanical (manual_clamp, collapsible if-let, doc-comment indentation, format-arg inlining, items_after_test_module relocation). Justified #[allow] retained for WinAPI type aliases (HICON/HDC/HGDIOBJ), a test helper's argument count, and deliberate const range-guard assertions.

+++ src-tauri/src/admin_autostart.rs
@@ -209,7 +209,7 @@ mod win {
         // BIND_OPTS3 -> BIND_OPTS2 -> BIND_OPTS is the layout windows-rs exposes.
         let mut bind_opts3 = BIND_OPTS3::default();
         bind_opts3.Base.Base.cbStruct = std::mem::size_of::<BIND_OPTS3>() as u32;
-        bind_opts3.Base.dwClassContext = CLSCTX_INPROC_SERVER.0 as u32;
+        bind_opts3.Base.dwClassContext = CLSCTX_INPROC_SERVER.0;
 
         let result: windows::core::Result<ITaskService> = unsafe {
             CoGetObject(
